requirements.txt
streamlit
google-generativeai
pdfminer.six
numpy
